+++ crates/content-base/src/db/search/mod.rs
@@ -377,7 +377,7 @@ mod test {
 
     #[test(tokio::test)]
     async fn test_select_text() {
-        let db = setup().await;
+        let db = setup(None).await;
         let text_res = db
             .select_text(vec!["text:7dd12x11yvt5fgamdjb0"])
             .await
@@ -387,7 +387,7 @@ mod test {
 
     #[test(tokio::test)]
     async fn test_select_image() {
-        let db = setup().await;
+        let db = setup(None).await;
         let image_res = db
             .select_image(vec!["image:flzkn6ncniglqttxnrsm"])
             .await
@@ -397,7 +397,7 @@ mod test {
 
     #[test(tokio::test)]
     async fn test_select_audio() {
-        let db = setup().await;
+        let db = setup(None).await;
         let audio_res = db
             .select_audio(vec!["audio:gkzq6db9jwr34l3j0gmz"])
             .await
@@ -407,7 +407,7 @@ mod test {
 
     #[test(tokio::test)]
     async fn test_select_video() {
-        let db = setup().await;
+        let db = setup(None).await;
         let video_res = db
             .select_video(vec!["video:u456grwuvl6w74zgqemc"])
             .await
@@ -417,7 +417,7 @@ mod test {
 
     #[test(tokio::test)]
     async fn test_select_web_page() {
-        let db = setup().await;
+        let db = setup(None).await;
         let web_page_res = db
             .select_web_page(vec!["web:nobc02c8ffyol3kqbsln"])
             .await
@@ -427,7 +427,7 @@ mod test {
 
     #[test(tokio::test)]
     async fn test_select_document() {
-        let db = setup().await;
+        let db = setup(None).await;
         let document_res = db
             .select_document(vec!["document:6dr6glzpf7ixefh7vjks"])
             .await
@@ -437,7 +437,7 @@ mod test {
 
     #[test(tokio::test)]
     async fn test_backtrace_by_ids() {
-        let db = setup().await;
+        let db = setup(None).await;
         let single_text_id = ID::from("text:11232131");
         db.upsert(&single_text_id, fake_upsert_text_clause().as_str())
             .await
@@ -483,7 +483,7 @@ mod test {
 
     #[test(tokio::test)]
     async fn test_full_text_search_with_highlight() {
-        let db = setup().await;
+        let db = setup(None).await;
         let res = db
             .full_text_search_with_highlight(vec!["LVL小河板".to_string()])
             .await
